+++ edge-nal-embassy/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+* Breaking: new trait: `UdpSplitMulticast` which is now required on the socket provided by `UdpBind` and `UdpConnect`
 * Update to `embassy-net` 0.9
 * Remove the unused `heapless` dependency
 * Raise MSRV to 1.91 to compile `smoltcp`